fix(freehand): rename the freehand-panel and the relative import statement
